+++ HPXMLtoOpenStudio/resources/hpxml.rb
@@ -51,7 +51,7 @@ require 'ostruct'
 class HPXML < Object
   HPXML_ATTRS = [:header, :site, :neighbor_buildings, :building_occupancy, :building_construction,
                  :climate_and_risk_zones, :air_infiltration_measurements, :attics, :foundations,
-                 :roofs, :rim_joists, :walls, :foundation_walls, :frame_floors, :slabs, :windows,
+                 :roofs, :rim_joists, :walls, :foundation_walls, :floors, :slabs, :windows,
                  :skylights, :doors, :partition_wall_mass, :furniture_mass, :heating_systems,
                  :cooling_systems, :heat_pumps, :hvac_plant, :hvac_controls, :hvac_distributions,
                  :ventilation_fans, :water_heating_systems, :hot_water_distributions, :water_fixtures,
@@ -119,8 +119,8 @@ class HPXML < Object
   FoundationWallTypeDoubleBrick = 'double brick'
   FoundationWallTypeSolidConcrete = 'solid concrete'
   FoundationWallTypeWood = 'wood'
-  FrameFloorOtherSpaceAbove = 'above'
-  FrameFloorOtherSpaceBelow = 'below'
+  FloorOtherSpaceAbove = 'above'
+  FloorOtherSpaceBelow = 'below'
   FuelLoadTypeGrill = 'grill'
   FuelLoadTypeLighting = 'lighting'
   FuelLoadTypeFireplace = 'fireplace'
@@ -434,7 +434,7 @@ class HPXML < Object
 
   def has_location(location)
     # Search for surfaces attached to this location
-    (@roofs + @rim_joists + @walls + @foundation_walls + @frame_floors + @slabs).each do |surface|
+    (@roofs + @rim_joists + @walls + @foundation_walls + @floors + @slabs).each do |surface|
       return true if surface.interior_adjacent_to == location
       return true if surface.exterior_adjacent_to == location
     end
@@ -582,7 +582,7 @@ class HPXML < Object
 
     # Get surfaces bounding infiltration volume
     spaces_within_infil_volume.each do |location|
-      (@roofs + @rim_joists + @walls + @foundation_walls + @frame_floors + @slabs).each do |surface|
+      (@roofs + @rim_joists + @walls + @foundation_walls + @floors + @slabs).each do |surface|
         is_adiabatic_surface = (surface.interior_adjacent_to == surface.exterior_adjacent_to)
         next unless [surface.interior_adjacent_to,
                      surface.exterior_adjacent_to].include? location
@@ -668,7 +668,7 @@ class HPXML < Object
     @rim_joists.to_oga(@doc)
     @walls.to_oga(@doc)
     @foundation_walls.to_oga(@doc)
-    @frame_floors.to_oga(@doc)
+    @floors.to_oga(@doc)
     @slabs.to_oga(@doc)
     @windows.to_oga(@doc)
     @skylights.to_oga(@doc)
@@ -722,7 +722,7 @@ class HPXML < Object
     @rim_joists = RimJoists.new(self, hpxml)
     @walls = Walls.new(self, hpxml)
     @foundation_walls = FoundationWalls.new(self, hpxml)
-    @frame_floors = FrameFloors.new(self, hpxml)
+    @floors = Floors.new(self, hpxml)
     @slabs = Slabs.new(self, hpxml)
     @windows = Windows.new(self, hpxml)
     @skylights = Skylights.new(self, hpxml)
@@ -886,7 +886,8 @@ class HPXML < Object
              :dst_enabled, :dst_begin_month, :dst_begin_day, :dst_end_month, :dst_end_day,
              :heat_pump_sizing_methodology, :allow_increased_fixed_capacities,
              :apply_ashrae140_assumptions, :energystar_calculation_version, :schedules_filepaths,
-             :occupancy_calculation_type, :extension_properties]
+             :occupancy_calculation_type, :extension_properties, :iecc_eri_calculation_version,
+             :zerh_calculation_version]
     attr_accessor(*ATTRS)
     attr_reader(:emissions_scenarios)
     attr_reader(:utility_bill_scenarios)
@@ -949,15 +950,15 @@ class HPXML < Object
         extension = XMLHelper.create_elements_as_needed(software_info, ['extension'])
         XMLHelper.add_element(extension, 'ApplyASHRAE140Assumptions', @apply_ashrae140_assumptions, :boolean) unless @apply_ashrae140_assumptions.nil?
       end
-      if not @eri_calculation_version.nil?
+      { @eri_calculation_version => 'ERICalculation',
+        @energystar_calculation_version => 'EnergyStarCalculation',
+        @iecc_eri_calculation_version => 'IECCERICalculation',
+        @zerh_calculation_version => 'ZERHCalculation' }.each do |calculation_version, element_name|
+        next if calculation_version.nil?
+
         extension = XMLHelper.create_elements_as_needed(software_info, ['extension'])
-        eri_calculation = XMLHelper.add_element(extension, 'ERICalculation')
-        XMLHelper.add_element(eri_calculation, 'Version', @eri_calculation_version, :string)
-      end
-      if not @energystar_calculation_version.nil?
-        extension = XMLHelper.create_elements_as_needed(software_info, ['extension'])
-        energystar_calculation = XMLHelper.add_element(extension, 'EnergyStarCalculation')
-        XMLHelper.add_element(energystar_calculation, 'Version', @energystar_calculation_version, :string)
+        calculation = XMLHelper.add_element(extension, element_name)
+        XMLHelper.add_element(calculation, 'Version', calculation_version, :string)
       end
       if (not @timestep.nil?) || (not @sim_begin_month.nil?) || (not @sim_begin_day.nil?) || (not @sim_end_month.nil?) || (not @sim_end_day.nil?) || (not @dst_enabled.nil?) || (not @dst_begin_month.nil?) || (not @dst_begin_day.nil?) || (not @dst_end_month.nil?) || (not @dst_end_day.nil?)
         extension = XMLHelper.create_elements_as_needed(software_info, ['extension'])
@@ -1037,7 +1038,9 @@ class HPXML < Object
       @software_program_used = XMLHelper.get_value(hpxml, 'SoftwareInfo/SoftwareProgramUsed', :string)
       @software_program_version = XMLHelper.get_value(hpxml, 'SoftwareInfo/SoftwareProgramVersion', :string)
       @eri_calculation_version = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/ERICalculation/Version', :string)
+      @iecc_eri_calculation_version = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/IECCERICalculation/Version', :string)
       @energystar_calculation_version = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/EnergyStarCalculation/Version', :string)
+      @zerh_calculation_version = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/ZERHCalculation/Version', :string)
       @timestep = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/SimulationControl/Timestep', :integer)
       @sim_begin_month = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/SimulationControl/BeginMonth', :integer)
       @sim_begin_day = XMLHelper.get_value(hpxml, 'SoftwareInfo/extension/SimulationControl/BeginDayOfMonth', :integer)
@@ -1582,7 +1585,7 @@ class HPXML < Object
 
   class Attic < BaseElement
     ATTRS = [:id, :attic_type, :vented_attic_sla, :vented_attic_ach, :within_infiltration_volume,
-             :attached_to_roof_idrefs, :attached_to_wall_idrefs, :attached_to_frame_floor_idrefs]
+             :attached_to_roof_idrefs, :attached_to_wall_idrefs, :attached_to_floor_idrefs]
     attr_accessor(*ATTRS)
 
     def attached_roofs
@@ -1607,12 +1610,12 @@ class HPXML < Object
       return list
     end
 
-    def attached_frame_floors
-      return [] if @attached_to_frame_floor_idrefs.nil?
+    def attached_floors
+      return [] if @attached_to_floor_idrefs.nil?
 
-      list = @hpxml_object.frame_floors.select { |frame_floor| @attached_to_frame_floor_idrefs.include? frame_floor.id }
-      if @attached_to_frame_floor_idrefs.size > list.size
-        fail "Attached frame floor not found for attic '#{@id}'."
+      list = @hpxml_object.floors.select { |floor| @attached_to_floor_idrefs.include? floor.id }
+      if @attached_to_floor_idrefs.size > list.size
+        fail "Attached floor not found for attic '#{@id}'."
       end
 
       return list
@@ -1640,7 +1643,7 @@ class HPXML < Object
       errors = []
       begin; attached_roofs; rescue StandardError => e; errors << e.message; end
       begin; attached_walls; rescue StandardError => e; errors << e.message; end
-      begin; attached_frame_floors; rescue StandardError => e; errors << e.message; end
+      begin; attached_floors; rescue StandardError => e; errors << e.message; end
       begin; to_location; rescue StandardError => e; errors << e.message; end
       return errors
     end
@@ -1691,9 +1694,9 @@ class HPXML < Object
           XMLHelper.add_attribute(wall_attached, 'idref', wall)
         end
       end
-      if not @attached_to_frame_floor_idrefs.nil?
-        @attached_to_frame_floor_idrefs.each do |floor|
-          floor_attached = XMLHelper.add_element(attic, 'AttachedToFrameFloor')
+      if not @attached_to_floor_idrefs.nil?
+        @attached_to_floor_idrefs.each do |floor|
+          floor_attached = XMLHelper.add_element(attic, 'AttachedToFloor')
           XMLHelper.add_attribute(floor_attached, 'idref', floor)
         end
       end
@@ -1729,9 +1732,9 @@ class HPXML < Object
       XMLHelper.get_elements(attic, 'AttachedToWall').each do |wall|
         @attached_to_wall_idrefs << HPXML::get_idref(wall)
       end
-      @attached_to_frame_floor_idrefs = []
-      XMLHelper.get_elements(attic, 'AttachedToFrameFloor').each do |frame_floor|
-        @attached_to_frame_floor_idrefs << HPXML::get_idref(frame_floor)
+      @attached_to_floor_idrefs = []
+      XMLHelper.get_elements(attic, 'AttachedToFloor').each do |floor|
+        @attached_to_floor_idrefs << HPXML::get_idref(floor)
       end
     end
   end
@@ -1752,7 +1755,7 @@ class HPXML < Object
 
   class Foundation < BaseElement
     ATTRS = [:id, :foundation_type, :vented_crawlspace_sla, :within_infiltration_volume,
-             :attached_to_slab_idrefs, :attached_to_frame_floor_idrefs, :attached_to_foundation_wall_idrefs,
+             :attached_to_slab_idrefs, :attached_to_floor_idrefs, :attached_to_foundation_wall_idrefs,
              :attached_to_wall_idrefs, :attached_to_rim_joist_idrefs]
     attr_accessor(*ATTRS)
 
@@ -1767,12 +1770,12 @@ class HPXML < Object
       return list
     end
 
-    def attached_frame_floors
-      return [] if @attached_to_frame_floor_idrefs.nil?
+    def attached_floors
+      return [] if @attached_to_floor_idrefs.nil?
 
-      list = @hpxml_object.frame_floors.select { |frame_floor| @attached_to_frame_floor_idrefs.include? frame_floor.id }
-      if @attached_to_frame_floor_idrefs.size > list.size
-        fail "Attached frame floor not found for foundation '#{@id}'."
+      list = @hpxml_object.floors.select { |floor| @attached_to_floor_idrefs.include? floor.id }
+      if @attached_to_floor_idrefs.size > list.size
+        fail "Attached floor not found for foundation '#{@id}'."
       end
 
       return list
@@ -1842,9 +1845,9 @@ class HPXML < Object
         sum_area += slab.area
       end
       if sum_area <= 0
-        # Check FrameFloors next
-        attached_frame_floors.each do |frame_floor|
-          sum_area += frame_floor.area
+        # Check Floors next
+        attached_floors.each do |floor|
+          sum_area += floor.area
         end
       end
       return sum_area
@@ -1857,7 +1860,7 @@ class HPXML < Object
     def check_for_errors
       errors = []
       begin; attached_slabs; rescue StandardError => e; errors << e.message; end
-      begin; attached_frame_floors; rescue StandardError => e; errors << e.message; end
+      begin; attached_floors; rescue StandardError => e; errors << e.message; end
       begin; attached_foundation_walls; rescue StandardError => e; errors << e.message; end
       begin; attached_walls; rescue StandardError => e; errors << e.message; end
       begin; attached_rim_joists; rescue StandardError => e; errors << e.message; end
@@ -1919,10 +1922,10 @@ class HPXML < Object
           XMLHelper.add_attribute(foundation_wall_attached, 'idref', foundation_wall)
         end
       end
-      if not @attached_to_frame_floor_idrefs.nil?
-        @attached_to_frame_floor_idrefs.each do |frame_floor|
-          floor_frame_attached = XMLHelper.add_element(foundation, 'AttachedToFrameFloor')
-          XMLHelper.add_attribute(floor_frame_attached, 'idref', frame_floor)
+      if not @attached_to_floor_idrefs.nil?
+        @attached_to_floor_idrefs.each do |floor|
+          floor_attached = XMLHelper.add_element(foundation, 'AttachedToFloor')
+          XMLHelper.add_attribute(floor_attached, 'idref', floor)
         end
       end
       if not @attached_to_slab_idrefs.nil?
@@ -1970,9 +1973,9 @@ class HPXML < Object
       XMLHelper.get_elements(foundation, 'AttachedToSlab').each do |slab|
         @attached_to_slab_idrefs << HPXML::get_idref(slab)
       end
-      @attached_to_frame_floor_idrefs = []
-      XMLHelper.get_elements(foundation, 'AttachedToFrameFloor').each do |frame_floor|
-        @attached_to_frame_floor_idrefs << HPXML::get_idref(frame_floor)
+      @attached_to_floor_idrefs = []
+      XMLHelper.get_elements(foundation, 'AttachedToFloor').each do |floor|
+        @attached_to_floor_idrefs << HPXML::get_idref(floor)
       end
       @attached_to_foundation_wall_idrefs = []
       XMLHelper.get_elements(foundation, 'AttachedToFoundationWall').each do |foundation_wall|
@@ -2599,21 +2602,21 @@ class HPXML < Object
     end
   end
 
-  class FrameFloors < BaseArrayElement
+  class Floors < BaseArrayElement
     def add(**kwargs)
-      self << FrameFloor.new(@hpxml_object, **kwargs)
+      self << Floor.new(@hpxml_object, **kwargs)
     end
 
     def from_oga(hpxml)
       return if hpxml.nil?
 
-      XMLHelper.get_elements(hpxml, 'Building/BuildingDetails/Enclosure/FrameFloors/FrameFloor').each do |frame_floor|
-        self << FrameFloor.new(@hpxml_object, frame_floor)
+      XMLHelper.get_elements(hpxml, 'Building/BuildingDetails/Enclosure/Floors/Floor').each do |floor|
+        self << Floor.new(@hpxml_object, floor)
       end
     end
   end
 
-  class FrameFloor < BaseElement
+  class Floor < BaseElement
     ATTRS = [:id, :exterior_adjacent_to, :interior_adjacent_to, :area, :insulation_id,
              :insulation_assembly_r_value, :insulation_cavity_r_value, :insulation_continuous_r_value,
              :other_space_above_or_below, :interior_finish_type, :interior_finish_thickness]
@@ -2624,7 +2627,7 @@ class HPXML < Object
         return true
       elsif [LocationAtticVented, LocationAtticUnvented].include? @exterior_adjacent_to
         return true
-      elsif [LocationOtherHousingUnit, LocationOtherHeatedSpace, LocationOtherMultifamilyBufferSpace, LocationOtherNonFreezingSpace].include?(@exterior_adjacent_to) && (@other_space_above_or_below == FrameFloorOtherSpaceAbove)
+      elsif [LocationOtherHousingUnit, LocationOtherHeatedSpace, LocationOtherMultifamilyBufferSpace, LocationOtherNonFreezingSpace].include?(@exterior_adjacent_to) && (@other_space_above_or_below == FloorOtherSpaceAbove)
         return true
       end
 
@@ -2664,15 +2667,15 @@ class HPXML < Object
     end
 
     def delete
-      @hpxml_object.frame_floors.delete(self)
+      @hpxml_object.floors.delete(self)
       @hpxml_object.attics.each do |attic|
-        attic.attached_to_frame_floor_idrefs.delete(@id) unless attic.attached_to_frame_floor_idrefs.nil?
+        attic.attached_to_floor_idrefs.delete(@id) unless attic.attached_to_floor_idrefs.nil?
       end
       @hpxml_object.foundations.each do |foundation|
-        foundation.attached_to_frame_floor_idrefs.delete(@id) unless foundation.attached_to_frame_floor_idrefs.nil?
+        foundation.attached_to_floor_idrefs.delete(@id) unless foundation.attached_to_floor_idrefs.nil?
       end
       @hpxml_object.attics.each do |attic|
-        attic.attached_to_frame_floor_idrefs.delete(@id) unless attic.attached_to_frame_floor_idrefs.nil?
+        attic.attached_to_floor_idrefs.delete(@id) unless attic.attached_to_floor_idrefs.nil?
       end
     end
 
@@ -2684,19 +2687,19 @@ class HPXML < Object
     def to_oga(doc)
       return if nil?
 
-      frame_floors = XMLHelper.create_elements_as_needed(doc, ['HPXML', 'Building', 'BuildingDetails', 'Enclosure', 'FrameFloors'])
-      frame_floor = XMLHelper.add_element(frame_floors, 'FrameFloor')
-      sys_id = XMLHelper.add_element(frame_floor, 'SystemIdentifier')
+      floors = XMLHelper.create_elements_as_needed(doc, ['HPXML', 'Building', 'BuildingDetails', 'Enclosure', 'Floors'])
+      floor = XMLHelper.add_element(floors, 'Floor')
+      sys_id = XMLHelper.add_element(floor, 'SystemIdentifier')
       XMLHelper.add_attribute(sys_id, 'id', @id)
-      XMLHelper.add_element(frame_floor, 'ExteriorAdjacentTo', @exterior_adjacent_to, :string) unless @exterior_adjacent_to.nil?
-      XMLHelper.add_element(frame_floor, 'InteriorAdjacentTo', @interior_adjacent_to, :string) unless @interior_adjacent_to.nil?
-      XMLHelper.add_element(frame_floor, 'Area', @area, :float) unless @area.nil?
+      XMLHelper.add_element(floor, 'ExteriorAdjacentTo', @exterior_adjacent_to, :string) unless @exterior_adjacent_to.nil?
+      XMLHelper.add_element(floor, 'InteriorAdjacentTo', @interior_adjacent_to, :string) unless @interior_adjacent_to.nil?
+      XMLHelper.add_element(floor, 'Area', @area, :float) unless @area.nil?
       if (not @interior_finish_type.nil?) || (not @interior_finish_thickness.nil?)
-        interior_finish = XMLHelper.add_element(frame_floor, 'InteriorFinish')
+        interior_finish = XMLHelper.add_element(floor, 'InteriorFinish')
         XMLHelper.add_element(interior_finish, 'Type', @interior_finish_type, :string, @interior_finish_type_isdefaulted) unless @interior_finish_type.nil?
         XMLHelper.add_element(interior_finish, 'Thickness', @interior_finish_thickness, :float, @interior_finish_thickness_isdefaulted) unless @interior_finish_thickness.nil?
       end
-      insulation = XMLHelper.add_element(frame_floor, 'Insulation')
+      insulation = XMLHelper.add_element(floor, 'Insulation')
       sys_id = XMLHelper.add_element(insulation, 'SystemIdentifier')
       if not @insulation_id.nil?
         XMLHelper.add_attribute(sys_id, 'id', @insulation_id)
@@ -2714,29 +2717,29 @@ class HPXML < Object
         XMLHelper.add_element(layer, 'InstallationType', 'continuous', :string)
         XMLHelper.add_element(layer, 'NominalRValue', @insulation_continuous_r_value, :float)
       end
-      XMLHelper.add_extension(frame_floor, 'OtherSpaceAboveOrBelow', @other_space_above_or_below, :string) unless @other_space_above_or_below.nil?
+      XMLHelper.add_extension(floor, 'OtherSpaceAboveOrBelow', @other_space_above_or_below, :string) unless @other_space_above_or_below.nil?
     end
 
-    def from_oga(frame_floor)
-      return if frame_floor.nil?
+    def from_oga(floor)
+      return if floor.nil?
 
-      @id = HPXML::get_id(frame_floor)
-      @exterior_adjacent_to = XMLHelper.get_value(frame_floor, 'ExteriorAdjacentTo', :string)
-      @interior_adjacent_to = XMLHelper.get_value(frame_floor, 'InteriorAdjacentTo', :string)
-      @area = XMLHelper.get_value(frame_floor, 'Area', :float)
-      interior_finish = XMLHelper.get_element(frame_floor, 'InteriorFinish')
+      @id = HPXML::get_id(floor)
+      @exterior_adjacent_to = XMLHelper.get_value(floor, 'ExteriorAdjacentTo', :string)
+      @interior_adjacent_to = XMLHelper.get_value(floor, 'InteriorAdjacentTo', :string)
+      @area = XMLHelper.get_value(floor, 'Area', :float)
+      interior_finish = XMLHelper.get_element(floor, 'InteriorFinish')
       if not interior_finish.nil?
         @interior_finish_type = XMLHelper.get_value(interior_finish, 'Type', :string)
         @interior_finish_thickness = XMLHelper.get_value(interior_finish, 'Thickness', :float)
       end
-      insulation = XMLHelper.get_element(frame_floor, 'Insulation')
+      insulation = XMLHelper.get_element(floor, 'Insulation')
       if not insulation.nil?
         @insulation_id = HPXML::get_id(insulation)
         @insulation_assembly_r_value = XMLHelper.get_value(insulation, 'AssemblyEffectiveRValue', :float)
         @insulation_cavity_r_value = XMLHelper.get_value(insulation, "Layer[InstallationType='cavity']/NominalRValue", :float)
         @insulation_continuous_r_value = XMLHelper.get_value(insulation, "Layer[InstallationType='continuous']/NominalRValue", :float)
       end
-      @other_space_above_or_below = XMLHelper.get_value(frame_floor, 'extension/OtherSpaceAboveOrBelow', :string)
+      @other_space_above_or_below = XMLHelper.get_value(floor, 'extension/OtherSpaceAboveOrBelow', :string)
     end
   end
 
@@ -6236,7 +6239,7 @@ class HPXML < Object
                    walls: @walls,
                    rim_joists: @rim_joists,
                    foundation_walls: @foundation_walls,
-                   frame_floors: @frame_floors,
+                   floors: @floors,
                    slabs: @slabs,
                    windows: @windows,
                    skylights: @skylights,
@@ -6305,7 +6308,7 @@ class HPXML < Object
   end
 
   def delete_tiny_surfaces()
-    (@rim_joists + @walls + @foundation_walls + @frame_floors + @roofs + @windows + @skylights + @doors + @slabs).reverse_each do |surface|
+    (@rim_joists + @walls + @foundation_walls + @floors + @roofs + @windows + @skylights + @doors + @slabs).reverse_each do |surface|
       next if surface.area.nil? || (surface.area > 1.0)
 
       surface.delete
@@ -6508,7 +6511,7 @@ class HPXML < Object
       return true
     elsif conditioned_locations.include?(surface.interior_adjacent_to) &&
           conditioned_locations.include?(surface.exterior_adjacent_to)
-      # E.g., frame floor between living space and conditioned basement, or
+      # E.g., floor between living space and conditioned basement, or
       # wall between living space and "other housing unit"
       return true
     end